Use SHA1 hash to cache Cythonized code
Previously setup.py used file timestamps to decide whether Cython
has to be executed again. This causes issues if e.g. changes happen
between different GIT branches etc.

This patch computes a SHA1 digest of _NetworKit.pyx and compares it
to the digest (if any) in _NetworKit.cpp. In case of match, Cython
does not have to be executed again.

In _NetworKit.cpp the digest is stored in the first line as C comment:
 "// source digest: ..."

Also inform the user, how to install Cython via pip.